requirement.txt
alembic==1.5.8
apache-airflow==2.0.2
apache-airflow-providers-ftp==1.0.1
apache-airflow-providers-http==1.1.1
apache-airflow-providers-imap==1.0.1
apache-airflow-providers-sqlite==1.0.2
apispec==3.3.2
argcomplete==1.12.3
asgiref==3.4.1
attrs==20.3.0
Babel==2.9.0
blinker==1.4
cached-property==1.5.2
cachetools==4.2.2
cattrs==1.0.0
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
connexion==2.7.0
croniter==0.3.37
cryptography==3.4.7
cycler==0.10.0
dataclasses==0.7
defusedxml==0.7.1
dill==0.3.2
dnspython==1.16.0
docutils==0.17.1
email-validator==1.1.2
et-xmlfile==1.1.0
fastapi==0.67.0
Flask==1.1.2
Flask-AppBuilder==3.2.3
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-Inputs==0.3.0
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-Pydantic==0.8.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
future==0.18.2
google-ads==4.0.0
google-api-core==1.31.0
google-api-python-client==1.12.8
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud==0.34.0
google-cloud-appengine-logging==0.1.1
google-cloud-audit-log==0.1.0
google-cloud-automl==2.3.0
google-cloud-bigquery==2.20.0
google-cloud-bigquery-datatransfer==3.2.0
google-cloud-bigquery-storage==2.5.0
google-cloud-bigtable==1.7.0
google-cloud-container==1.0.1
google-cloud-core==1.7.1
google-cloud-datacatalog==3.2.1
google-cloud-dataproc==2.4.0
google-cloud-dlp==1.0.0
google-cloud-kms==2.3.0
google-cloud-language==1.3.0
google-cloud-logging==2.5.0
google-cloud-memcache==1.0.0
google-cloud-monitoring==2.4.0
google-cloud-os-login==2.2.1
google-cloud-pubsub==2.6.1
google-cloud-redis==2.1.1
google-cloud-secret-manager==1.0.0
google-cloud-spanner==1.19.1
google-cloud-speech==1.3.2
google-cloud-storage==1.40.0
google-cloud-tasks==2.4.0
google-cloud-texttospeech==1.0.1
google-cloud-translate==1.7.0
google-cloud-videointelligence==1.16.1
google-cloud-vision==1.0.0
google-cloud-workflows==1.1.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
graphviz==0.16
grpc-google-iam-v1==0.12.3
grpcio==1.38.1
grpcio-gcp==0.2.2
gunicorn==19.10.0
h11==0.12.0
httplib2==0.19.1
idna==2.10
importlib-metadata==1.7.0
importlib-resources==1.5.0
inflection==0.5.1
iso8601==0.1.14
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
json-merge-patch==0.2
jsonschema==3.2.0
kafka-python==2.0.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
libcst==0.3.19
lockfile==0.12.2
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==2.1.0
marshmallow-sqlalchemy==0.23.1
matplotlib==3.3.4
mypy-extensions==0.4.3
natsort==7.1.1
numpy==1.19.5
oauthlib==3.1.1
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.0
openpyxl==3.0.7
packaging==20.9
pandas==1.1.5
pandas-gbq==0.14.1
pendulum==2.1.2
pep562==1.0
Pillow==8.2.0
pkg-resources==0.0.0
plyvel==1.3.0
prison==0.1.3
proto-plus==1.19.0
protobuf==3.17.3
psutil==5.8.0
psycopg2==2.9.1
pubsub==0.1.2
py4j==0.10.9
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pydata-google-auth==1.2.0
Pygments==2.8.1
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pyspark==3.1.2
python-daemon==2.3.0
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rich==9.2.0
rsa==4.7.2
scipy==1.5.4
seaborn==0.11.1
setproctitle==1.2.2
six==1.15.0
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.0
starlette==0.14.2
swagger-ui-bundle==0.0.8
tabulate==0.8.9
tenacity==6.2.0
termcolor==1.1.0
text-unidecode==1.3
typing==3.7.4.3
typing-extensions==3.7.4.3
typing-inspect==0.7.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.11
uvicorn==0.14.0
Werkzeug==1.0.1
WTForms==2.3.3
xlrd==2.0.1
zipp==3.4.1

